backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-community==0.0.10
faiss-cpu==1.7.4
sentence-transformers==2.7.0
feedparser==6.0.10
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
openai==1.3.0
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
numpy==1.24.3
pandas==2.1.4
python-multipart==0.0.6
huggingface_hub==0.20.3
